+++ dcmweb/requests_util.py
@@ -7,19 +7,10 @@ import os
 from threading import Lock
 import urllib.parse as urlparse
 import requests
-import validators
 
+from . import resources
 
 PAGE_SIZE = 5000
-
-STUDY_TAG = "0020000D"
-SERIES_TAG = "0020000E"
-INSTANCE_TAG = "00080018"
-
-STUDY_ID = "study_id"
-SERIES_ID = "series_id"
-INSTANCE_ID = "instance_id"
-FRAME_ID = "frame_id"
 
 DCM_EXTENSION = ".dcm"
 JPEG_EXTENSION = ".jpg"
@@ -27,11 +18,6 @@ PNG_EXTENSION = ".png"
 
 CONTENT_TYPE = "Content-Type"
 MULTIPART = "multipart/related"
-
-ID_PATH_MAP = {STUDY_ID: "studies", SERIES_ID: "series",
-               INSTANCE_ID: "instances", FRAME_ID: "frames"}
-
-SPLIT_CHAR = '/'
 
 
 def filter_urllib3_logging():
@@ -56,50 +42,6 @@ class NoHeaderErrorFilter(logging.Filter):  # pylint: disable=too-few-public-met
 filter_urllib3_logging()
 
 
-class NetworkError(Exception):
-    """exception for unexpected responses"""
-
-
-def validate_host_str(host):
-    """Function to check host url"""
-    if not validators.url(host):
-        raise ValueError('Invalid url')
-    if host[-1] != SPLIT_CHAR:
-        host += SPLIT_CHAR
-    return host
-
-
-def validate_path(path):
-    """Function to check path"""
-    if path:
-        if path[0] == SPLIT_CHAR:
-            path = path[1:]
-        if path[-1] == SPLIT_CHAR:
-            path = path[:-1]
-        path_splitted = path.split(SPLIT_CHAR)
-        number_of_pieces = len(path_splitted)
-        if number_of_pieces < 2 or number_of_pieces % 2 \
-                or path_splitted[-2] not in ["root", "studies", "series", "instances", "frames"]:
-            raise ValueError("incorrect path")
-    return path
-
-
-def get_dicom_tag(dictionary, tag):
-    """Wrapper for dicom json dict"""
-    if tag not in dictionary:
-        raise LookupError("Can't find {} tag ".format(tag))
-    return dictionary[tag]["Value"][0]
-
-
-def ids_from_json(json_dict):
-    """Generates dict of ids based on json dict object"""
-    ids = {}
-    ids[STUDY_ID] = get_dicom_tag(json_dict, STUDY_TAG)
-    ids[SERIES_ID] = get_dicom_tag(json_dict, SERIES_TAG)
-    ids[INSTANCE_ID] = get_dicom_tag(json_dict, INSTANCE_TAG)
-    return ids
-
-
 def add_limit_if_not_present(parameters, limit=PAGE_SIZE):
     """Adds limit parameter if it's not present"""
     if "limit" not in parameters:
@@ -107,75 +49,6 @@ def add_limit_if_not_present(parameters, limit=PAGE_SIZE):
             parameters += "&"
         parameters += "limit={}".format(limit)
     return parameters
-
-
-def get_path_level(ids):
-    """Return level of path
-    :param: a dict of ids study_id:<uid>, series_id:<uid>,\
-              instance_id:<uid>, [frame_id:<frame_num>]
-    :returns: string level(root, studies, series, instances, frames)"""
-    if not ids:
-        return "root"
-    if FRAME_ID in ids:
-        return ID_PATH_MAP[FRAME_ID]
-    if INSTANCE_ID in ids:
-        return ID_PATH_MAP[INSTANCE_ID]
-    if SERIES_ID in ids:
-        return ID_PATH_MAP[SERIES_ID]
-    if STUDY_ID in ids:
-        return ID_PATH_MAP[STUDY_ID]
-    raise ValueError("unknown level of path")
-
-
-def ids_from_path(path):
-    """Parses path to get Ids of study, series, instance and frame as optional
-    :param path: path to dicom object(instance or frame)
-             in form of url path /studies/<uid>/series/<uid>/instances/<uid>[/frames/<frame_num>]
-    :returns: a dict of ids study_id:<uid>, series_id:<uid>,\
-              instance_id:<uid>, [frame_id:<frame_num>]
-    """
-    path = validate_path(path)
-    path_splitted = path.split(SPLIT_CHAR)
-    ids = {}
-
-    if len(path_splitted) >= 2:
-        ids[STUDY_ID] = path_splitted[1]
-
-    if len(path_splitted) >= 4:
-        ids[SERIES_ID] = path_splitted[3]
-
-    if len(path_splitted) >= 6:
-        ids[INSTANCE_ID] = path_splitted[5]
-
-    if len(path_splitted) >= 8:
-        ids[FRAME_ID] = path_splitted[7]
-    return ids
-
-
-def path_from_ids(ids):
-    """Builds path based on dict of ids
-    :returns: a dict of ids study_id:<uid>, series_id:<uid>,\
-              instance_id:<uid>, [frame_id:<frame_num>]
-    :param path: path to dicom object(instance or frame)
-             in form of url path /studies/<uid>/series/<uid>/instances/<uid>[/frames/<frame_num>]
-    """
-    path = ""
-    if not ids:
-        return path
-    if STUDY_ID in ids:
-        path += id_to_string(STUDY_ID, ids[STUDY_ID])
-    if SERIES_ID in ids:
-        path += id_to_string(SERIES_ID, ids[SERIES_ID])
-    if INSTANCE_ID in ids:
-        path += id_to_string(INSTANCE_ID, ids[INSTANCE_ID])
-    if FRAME_ID in ids:
-        path += id_to_string(FRAME_ID, ids[FRAME_ID])
-    return path
-
-
-def id_to_string(id_key, id_value):
-    """ Builds string based kay and value of id"""
-    return SPLIT_CHAR+ID_PATH_MAP[id_key]+SPLIT_CHAR+id_value
 
 
 def extention_by_headers(content_type):
@@ -190,28 +63,23 @@ def extention_by_headers(content_type):
     raise ValueError("unknown extention {}".format(content_type))
 
 
-def file_system_full_path_by_ids(ids, base_dir="./"):
-    """Builds file system path and file name based on ids"""
-    path = ids[STUDY_ID] + SPLIT_CHAR + ids[SERIES_ID] + SPLIT_CHAR
-    if base_dir[-1] != SPLIT_CHAR:
-        path = SPLIT_CHAR + path
-    path = base_dir + path
-    file_name = ids[INSTANCE_ID]
-    return path, file_name
-
-
 def parse_boundary(content_type):
     """Returns boundary from content type"""
     boundary_start = content_type.find("boundary=")+9
     return bytes(content_type[boundary_start:content_type.find(
         ";", boundary_start)], "utf-8")
 
+
+class NetworkError(Exception):
+    """exception for unexpected responses"""
+
+
 class Requests:
     """Class keep state of credentials
      and performs request to dicomWeb"""
 
     def __init__(self, host_str, authenticator):
-        self.host = validate_host_str(host_str)
+        self.host = resources.validate_host_str(host_str)
         self.authenticator = authenticator
         self.authenticator_lock = Lock()
 
@@ -250,7 +118,7 @@ class Requests:
 
     def delete_dicom(self, path):
         """ Deletes single dicom object by sending DELETE http request"""
-        path = validate_path(path)
+        path = resources.validate_path(path)
         response = requests.delete(self.build_url(
             path, ""), headers=self.apply_credentials({}))
         if response.status_code != 200:
@@ -266,10 +134,11 @@ class Requests:
         if "limit" in par:
             limit = int(par["limit"][0])
         if limit > PAGE_SIZE:
-            raise ValueError("limit can\'t be more than {}".format(PAGE_SIZE))
+            raise ValueError("limit can\'t be more than {}".format(
+                PAGE_SIZE))
 
         text = self.request(
-            path_from_ids(
+            resources.path_from_ids(
                 ids)+"/instances", add_limit_if_not_present(parameters, limit)
             + "&offset={}".format(limit*page), {}).text
         return text
@@ -322,18 +191,19 @@ class Requests:
 
     def download_dicom_by_ids(self, ids, output="./", mime_type=None):
         """Downloads instance based on ids dict object"""
-        url = path_from_ids(ids)
-        folder, file_name = file_system_full_path_by_ids(ids, output)
+        url = resources.path_from_ids(ids)
+        folder, file_name = resources.file_system_full_path_by_ids(ids, output)
         return self.download_dicom(url, folder, file_name, mime_type)
 
     def build_url(self, path, parameters):
         """Builds url from host and path"""
         path_str = str(path)
-        if len(path_str) > 0 and path_str[0] == SPLIT_CHAR:
+        if len(path_str) > 0 and path_str[0] == resources.SPLIT_CHAR:
             path_str = path_str[1:]
         if parameters and parameters[0] != '?':
             path_str += "?"
         return self.host+path_str+parameters
+
 
 class MultipartChunksReader:  # pylint: disable=too-few-public-methods; need for readability
     """Class keep state of multipart stream"""
